Move live process into front end; update some tests; fix undo paste logic
